requirements.txt
python-docx>=1.2.0
selenium
webdriver-manager
beautifulsoup4
lxml
fastapi
uvicorn[standard]
pydantic
python-dotenv
pypdf
numpy
scikit-learn
joblib
requests
striprtf
sentence-transformers
torch
pytest
httpx
